+++ hadoop-hdfs-project/hadoop-hdfs/src/test/java/org/apache/hadoop/hdfs/util/TestStripedBlockUtil.java
@@ -126,9 +126,9 @@ public class TestStripedBlockUtil {
     DatanodeInfo[] locs = new DatanodeInfo[BLK_GROUP_WIDTH];
     String[] storageIDs = new String[BLK_GROUP_WIDTH];
     StorageType[] storageTypes = new StorageType[BLK_GROUP_WIDTH];
-    int[] indices = new int[BLK_GROUP_WIDTH];
+    byte[] indices = new byte[BLK_GROUP_WIDTH];
     for (int i = 0; i < BLK_GROUP_WIDTH; i++) {
-      indices[i] = (i + 2) % DATA_BLK_NUM;
+      indices[i] = (byte) ((i + 2) % DATA_BLK_NUM);
       // Location port always equal to logical index of a block,
       // for easier verification
       locs[i] = DFSTestUtil.getLocalDatanodeInfo(indices[i]);
